JP Onboarding: Move Summary Completed Steps to Separate Component
